+++ assets/js/admin/jsModule.js
@@ -1,0 +1,5 @@
+const init = () => {
+
+}
+
+export default init