+++ assets/js/theme/slideshow.js
@@ -1,0 +1,8 @@
+import UIkit from './UIKit'
+
+const initSlideshows = () => {
+  const slideshows = UIkit
+    .slideshow('.uk-slideshow', {})
+}
+
+export default initSlideshows