+++ sidebar.php
@@ -9,6 +9,6 @@ if (! is_active_sidebar('primary')) {
 }
 ?>
 
-<aside id="secondary" class="widget-area col-sm-4" role="complementary">
+<aside id="secondary" class="widget-area uk-width-1-4@s" role="complementary">
     <?php dynamic_sidebar('primary'); ?>
 </aside><!-- #secondary -->