+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [3.0.0] - 2022-03-29
+
+### Changed
+- Switched build process to use Rollup.js to support ES modules rather than modular pattern
+- Remove Bootstrap
+- Add UIKit for front end components
+- Build processes are now moved to a separate folder
+- Updated dependencies
+- Support Composer 2
+- Require PHP >= 7.4
+
 ## [2.4.0] - 2019-03-29
 ### Changed
 - Updated Gulp to v4